Backend/requirements.txt
amqp
asgiref
async-timeout
billiard
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cryptography
defusedxml
dj-database-url
Django==4.1.3
django-cors-headers
django-filter
django-redis
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
idna
inflection
kombu
Markdown
oauthlib
packaging
prompt-toolkit
psycopg2-binary
pycparser
PyJWT
python-dateutil
python3-openid
pytz
PyYAML
redis
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
whitenoise
